initial shell of new version of the server connect modal
