+++ js/templates/serverConnectHeader.html
@@ -1,0 +1,2 @@
+<div class="modal-hero-icon"></div>
+<p class="h6 homeModal-heroText fancy-heading <%= ob.msgClass || '' %>"><%= ob.msg %></p>
